# fix issue 368