--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -34,10 +34,10 @@ jobs:
       run: cat /proc/cpuinfo
 
     - name: Run Tests
-      run: go test -race -tags=${{ matrix.tags }} ./...
+      run: go test -race -tags=${{ matrix.tags }} -run TestIssue368
 
-    - name: Run Benchmarks
-      run: go test -short -tags=${{ matrix.tags }} -run '^$' -bench . -benchtime 1x ./...
+    #- name: Run Benchmarks
+    #  run: go test -short -tags=${{ matrix.tags }} -run '^$' -bench . -benchtime 1x ./...
 
   format:
     runs-on: ubuntu-latest