--- a/sparse/gather.go
+++ b/sparse/gather.go
@@ -26,7 +26,15 @@ func GatherUint64(dst []uint64, src Uint64Array) int { return gather64(dst, src)
 
 func GatherUint128(dst [][16]byte, src Uint128Array) int { return gather128(dst, src) }
 
-func GatherString(dst []string, src StringArray) int { return gatherString(dst, src) }
+func GatherString(dst []string, src StringArray) int {
+	n := min(len(dst), src.Len())
+
+	for i := range dst[:n] {
+		dst[i] = src.Index(i)
+	}
+
+	return n
+}
 
 func min(a, b int) int {
 	if a < b {