--- a/sparse/gather_amd64.go
+++ b/sparse/gather_amd64.go
@@ -3,8 +3,6 @@
 package sparse
 
 import (
-	"unsafe"
-
 	"golang.org/x/sys/cpu"
 )
 
@@ -85,11 +83,3 @@ func gather64AVX2(dst []uint64, src Uint64Array)
 
 //go:noescape
 func gather128(dst [][16]byte, src Uint128Array) int
-
-// On AMD64 the `string` type is made of a 64 bit pointer and a 64 bit integer,
-// for a total of 128 bit which means that we can use the same algorithm as
-// gather128 to lookup string values, since the memory layout of the arguments is
-// the same.
-func gatherString(dst []string, src StringArray) int {
-	return gather128(*(*[][16]byte)(unsafe.Pointer(&dst)), *(*Uint128Array)(unsafe.Pointer(&src)))
-}