--- a/sparse/gather_purego.go
+++ b/sparse/gather_purego.go
@@ -70,13 +70,3 @@ func gather128(dst [][16]byte, src Uint128Array) int {
 
 	return n
 }
-
-func gatherString(dst []string, src StringArray) int {
-	n := min(len(dst), src.Len())
-
-	for i := range dst[:n] {
-		dst[i] = src.Index(i)
-	}
-
-	return n
-}