… testdata/issue368.parquet