Fixes #368 

Pulls the file from https://github.com/segmentio/parquet-go/pull/373

The _fix_ results in a performance regression on lookups into byte array dictionaries:
```
name                                    old time/op  new time/op  delta
Dictionary/Lookup/BYTE_ARRAY/N=100       106ns ± 1%   253ns ± 1%  +137.77%  (p=0.000 n=10+10)
Dictionary/Lookup/BYTE_ARRAY/N=1000      841ns ± 0%  2335ns ± 0%  +177.53%  (p=0.000 n=9+10)
Dictionary/Lookup/BYTE_ARRAY/N=10000    10.4µs ± 0%  25.0µs ± 0%  +141.28%  (p=0.000 n=9+10)
Dictionary/Lookup/BYTE_ARRAY/N=100000    188µs ± 4%   297µs ± 1%   +58.22%  (p=0.000 n=10+9)
Dictionary/Lookup/BYTE_ARRAY/N=1000000  3.59ms ±11%  4.47ms ± 3%   +24.72%  (p=0.000 n=9+10)

name                                    old value/s  new value/s  delta
Dictionary/Lookup/BYTE_ARRAY/N=100        940M ± 1%    395M ± 1%   -57.94%  (p=0.000 n=10+10)
Dictionary/Lookup/BYTE_ARRAY/N=1000      1.19G ± 0%   0.43G ± 0%   -63.97%  (p=0.000 n=9+10)
Dictionary/Lookup/BYTE_ARRAY/N=10000      963M ± 0%    399M ± 0%   -58.55%  (p=0.000 n=9+10)
Dictionary/Lookup/BYTE_ARRAY/N=100000     533M ± 4%    336M ± 1%   -36.82%  (p=0.000 n=10+9)
Dictionary/Lookup/BYTE_ARRAY/N=1000000    280M ±12%    224M ± 3%   -20.16%  (p=0.000 n=9+10)
```